배경택: [Add] RecordManager 페이지 및 도감아이템 개수에 맞게 최초 생성 구성완료
