+++ Metalord/Assets/Resources/Prefabs/Object_ForRecordObject.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &2871829991851057285
+--- !u!1 &5315895649336509587
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,24 +8,24 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 6125091339123743652}
-  - component: {fileID: 2067380541515630845}
-  - component: {fileID: 8671927041378270438}
-  - component: {fileID: 6459172115803583224}
+  - component: {fileID: 7625121281377396403}
+  - component: {fileID: 446096968753651037}
+  - component: {fileID: 4196610910418584169}
+  - component: {fileID: 8097983158804455907}
   m_Layer: 5
-  m_Name: Object_forRecord
+  m_Name: Object_ForRecordObject
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!224 &6125091339123743652
+--- !u!224 &7625121281377396403
 RectTransform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2871829991851057285}
+  m_GameObject: {fileID: 5315895649336509587}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -33,26 +33,26 @@ RectTransform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 0, y: 0}
+  m_SizeDelta: {x: 100, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!222 &2067380541515630845
+--- !u!222 &446096968753651037
 CanvasRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2871829991851057285}
+  m_GameObject: {fileID: 5315895649336509587}
   m_CullTransparentMesh: 1
---- !u!114 &8671927041378270438
+--- !u!114 &4196610910418584169
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2871829991851057285}
+  m_GameObject: {fileID: 5315895649336509587}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
@@ -76,19 +76,15 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &6459172115803583224
+--- !u!114 &8097983158804455907
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2871829991851057285}
+  m_GameObject: {fileID: 5315895649336509587}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 76d3598795e024c5a89a61a31e235406, type: 3}
+  m_Script: {fileID: 11500000, guid: 1fb189dd20a604b1eb1acc3f8cc89ced, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  objectID: 
-  objectPlace: 
-  objectName: 
-  objectDescription:
