+++ Metalord/Assets/Resources/Prefabs/Object_ForRecordObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cfcca3ccc25a4478db199820751739cb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
