+++ Metalord/Assets/_Test/BKT/Scripts/Managers.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 5fbe7faca1eca405c82993f02eeaf3b5
-PrefabImporter:
+guid: 685e7dc32ff6b4e81808d9658a9c6512
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
