+++ Metalord/Assets/_Test/BKT/Scripts/Record.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10544d47cd75940218ef9141951f9b7a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
