+++ Metalord/Assets/_Test/BKT/Scripts/Record/RecordObject.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RecordObject : MonoBehaviour
+{
+    public RecordObjectInfo recordInfo { get; set; }
+
+
+
+    private void ReflectInfo()
+    {
+
+    }
+}
