+++ Metalord/Assets/_Test/BKT/Scripts/Record/RecordObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1fb189dd20a604b1eb1acc3f8cc89ced
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
