+++ Metalord/Assets/_Test/BKT/Scripts/Record/RecordObjectInfo.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 /// 도감 오브젝트의 정보_CSV를 통해 반영
 /// 231130_배경택
 /// </summary>
-public class ObjectInfo : MonoBehaviour
+public class RecordObjectInfo
 {
     public int id { get; set; }
     public string id_Description { get; set; }
@@ -14,6 +14,4 @@ public class ObjectInfo : MonoBehaviour
     public int zone { get; set; } // 1 = 주방, 2= 거실, 3= 아기방
     public bool obtained { get; set; }
     public string description { get; set; }
-
-    
 }
